// frontend/src/types/database.types.ts
export type LotStatus =
  | "publie"
  | "reserve"
  | "collecte"
  | "livre_recycleur"
  | "traite";

export type TypeDechet =
  | "plastique"
  | "metal"
  | "papier_carton"
  | "verre"
  | "organique"
  | "electronique"
  | "textile"
  | "mixte"
  | "inconnu";
// ...
export type LotRow = {
  id: string;
  producteur_id: string;
  type_dechet: TypeDechet;
  status: LotStatus;
  score_tri: number | null;
  volume_ia: number | null;
  weight_real: number | null;
  latitude: number | null;
  longitude: number | null;
  adresse_texte: string | null;
  commune: string | null;
  quartier: string | null;
  photo_url: string | null;
  disponibilite: string | null;
  note: string | null;
  collecteur_id: string | null;
  date_publication: string;
  date_collecte: string | null;
  date_livraison: string | null;
};